use INT_MAX or INT32_MAX when assigning to int or int32, instead of LONG_MAX
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2609 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_dialog_widgets_sdl.cpp
@@ -809,10 +809,10 @@ w_players_in_game2::draw_player_names_clumped(SDL_Surface* s, TextLayoutHelper& 
 
 int
 w_players_in_game2::find_maximum_bar_value() const {
-    int	theMaxValue = LONG_MIN;
+    int	theMaxValue = INT_MIN;
 
     // We track min also to handle games with negative scores.
-    int theMinValue = LONG_MAX;
+    int theMinValue = INT_MAX;
 
     if(selected_player != NONE)
         // This way, all player vs player graphs use the same scale.
